avoid overflow on 32 bit
Fix Data.BloomFilter.Easy on 32 bit to not incorrectly fail with "capacity too
large to represent".

Debian and NixOS have been reverting 44b01ba38b4fcdb5a85f44fa2f3af1f29cde8f40
for years because of this problem.

A too large value on 32 bit will overflow on its own, before it reaches
0xffffffff, because a signed int maxes out at 0x7fffffff.
So the roundedBits <= 0 check will catch it, and the other check can
be avoided on 32 bit.